Split preset storage by platform
